ARROW-9264: [C++][Parquet] Refactor and modernize schema conversion code
Split this off from ARROW-8493 because enough code moved around to
warrant a review without other functional changes.

- Rename max_ to either current_ or removed entirely
- During SchemaField generation remove one check for grouped
  child fields.
- Add a field that will be required for generalized resolution
  ARROW-8493
